Create README.md
This chat fully works on JavaScript with backend on Node.js and uses Socket.IO library. It has ability of authorization and allows to send txt messages, photos, files of any format, emoji, delete messages and also it has automatic scroll bar panel with automatic loading of messages. Chat was slightly modified for hackaton and an additional video bot was included in it.